+++ audioreception/app.py
@@ -52,16 +52,18 @@ def upload(): #uploaded soundblob from js client
     xfilename = request.headers.get('x-filename') # necessary?
     xclipid = request.headers.get('x-clipid')
     xtension = request.headers.get('x-tension') # .ogg, .mp4, .webm validation?
+    xsurveyid = request.headers.get('x-surveyid')
+
     xresponseid = request.headers.get('x-responseid')
 
-    # credate folder with responseid
-    if exists(receptiondir + xresponseid):
+    # credate folder with surveyid / responseid
+    if exists(receptiondir + xsurveyid + "/" + xresponseid):
         app.logger.info('it does exist', receptiondir)
     else:
         app.logger.info('it does NOT exist')
-        makedirs(receptiondir + xresponseid)
+        makedirs(receptiondir + xsurveyid + "/" + xresponseid)
 
-    responsefolder = receptiondir + xresponseid + "/"
+    responsefolder = receptiondir + xsurveyid + "/" + xresponseid + "/"
 
     ct = datetime.now()
     currentTime = ct.strftime("-%Y-%m-%d-%H-%M-%S.%f")
@@ -83,19 +85,52 @@ def upload(): #uploaded soundblob from js client
     else:
         return 'geen fetch'    
 
+
+# TODO rewrite it for the new structure
 @app.route('/watch/')
 @app.route('/watch/reception/')
-@app.route('/watch/<typewatch>/<respid>/')
-def watch(typewatch = None, respid = None):
-    if typewatch == "reception" and respid is not None and exists(receptiondir + respid):      
-        dir = "static/storage/reception/" + respid
-        lijst = listFiles(dir)
-        return render_template("index.html", lijst=lijst, dir=dir )
+@app.route('/watch/reception/<surveyid>/')
+@app.route('/watch/reception/<surveyid>/<responseid>/')
+
+# @app.route('/watch/<typewatch>/<surveyid>/')
+def watch(surveyid = None, responseid = None):
+    # app.logger.info('surveyid ' + surveyid + "/")
+
+    if (responseid is not None and surveyid is not None and exists("static/storage/reception/" + surveyid + "/" + responseid + "/")):
+        app.logger.info('j', responseid)
+        lijst = listDirs("static/storage/reception/" + surveyid + "/" + responseid + "/")
+
+        
+        return render_template("listresponses.html", lijst=lijst, dir="static/storage/reception/" + surveyid + "/" + responseid + "/")
+
+    elif (surveyid is not None and exists("static/storage/reception/" + surveyid + "/") ):
+        app.logger.info('j', surveyid)
+        lijst = listDirs("static/storage/reception/" + surveyid + "/")
+        return render_template("listresponses.html", lijst=lijst, dir="static/storage/reception/" + surveyid + "/" )
+        return 'hey: ' + surveyid + ", ".join(lijst)
     else:
-        lijst = listDirs(receptiondir)
-        return render_template("listresponses.html", lijst=lijst, dir=receptiondir )
+        lijst = listDirs("static/storage/reception/")
+        return render_template("listresponses.html", lijst=lijst, dir="static/storage/reception/")
 
+        return 'Nothing'    
 
+    # if exists(static/storage/reception/ + "/" + surveyid + "/"):    
+    #     app.logger.info('j')
+    #     dir = "static/storage/reception/" + surveyid
+    #     lijst = listFiles(dir)
+    #     print(lijst)
+    #     return render_template("index.html", lijst=lijst, dir=dir )
+    # else:
+    #     app.logger.info('n')
+
+    #     lijst = listDirs(receptiondir)
+    #     app.logger.info(lijst)
+
+    #     return render_template("listresponses.html", lijst=lijst, dir=receptiondir )
+
+@app.route('/test/')
+def test():
+    return 'hoi'
 
 if __name__ == "__main__":
     app.run(debug=True)